refactor: extract escapeHtml to shared utility module
Move the duplicated escapeHtml function from both src/markdown.ts and
src/tool-response-format.ts into a shared src/escape-html.ts module.
Both files now import from the single source of truth, keeping escaping
behavior consistent across the codebase.
